make new api the default
