+++ packages/vulcan-core/lib/modules/containers/delete.js
@@ -33,13 +33,13 @@ import { deleteClientTemplate } from 'meteor/vulcan:core';
 import { extractCollectionInfo, extractFragmentInfo } from 'meteor/vulcan:lib';
 import { buildMultiQuery } from './multi';
 import { getVariablesListFromCache, removeFromData } from './cacheUpdate';
+import { computeQueryVariables } from './variables';
 
-export const buildDeleteQuery = ({ typeName, fragmentName, fragment }) => (
+export const buildDeleteQuery = ({ typeName, fragmentName, fragment }) =>
   gql`
     ${deleteClientTemplate({ typeName, fragmentName })}
     ${fragment}
-  `
-);
+  `;
 
 // remove value from the cached lists
 const multiQueryUpdater = ({ collection, typeName, fragmentName, fragment }) => {
@@ -65,48 +65,50 @@ const multiQueryUpdater = ({ collection, typeName, fragmentName, fragment }) => 
   };
 };
 
-export const useDelete = (options) => {
+export const useDelete = options => {
   const { collectionName, collection } = extractCollectionInfo(options);
   const { fragmentName, fragment } = extractFragmentInfo(options, collectionName);
   const typeName = collection.options.typeName;
-  const resolverName = `delete${typeName}`;
-  const { mutationOptions = {} } = options;
+  const {
+    //input: optionsInput,
+    //_id: optionsId,
+    mutationOptions = {},
+  } = options;
 
   const query = buildDeleteQuery({
     fragment,
-    fragmentName, typeName
+    fragmentName,
+    typeName,
   });
 
   const [deleteFunc, ...rest] = useMutation(query, {
     // optimistic update
     update: multiQueryUpdater({ collection, typeName, fragment, fragmentName }),
-    ...mutationOptions
+    ...mutationOptions,
   });
-
-  const extendedDeleteFunc = {
-    [resolverName]: (args) => {
-      // support legacy syntax mistake
-      // @see https://github.com/VulcanJS/Vulcan/issues/2417
-      const selector = (args && args.selector) || args;
-      return deleteFunc({ variables: { selector } });
-    }
-  }
-  return [extendedDeleteFunc[resolverName], ...rest];
+  const extendedDeleteFunc = (args /*{ input: argsInput, _id: argsId }*/) => {
+    return deleteFunc({
+      variables: {
+        ...computeQueryVariables(options, args),
+      },
+    });
+  };
+  return [extendedDeleteFunc, ...rest];
 };
 
 export const withDelete = options => C => {
   const { collection } = extractCollectionInfo(options);
   const typeName = collection.options.typeName;
   const funcName = `delete${typeName}`;
+  const metaName = `delete${typeName}Meta`;
+
   const legacyError = () => {
     throw new Error(`removeMutation function has been removed. Use ${funcName} function instead.`);
   };
 
-  const Wrapper = (props) => {
-    const [deleteFunc] = useDelete(options);
-    return (
-      <C {...props} {...{ [funcName]: deleteFunc }} removeMutation={legacyError} />
-    );
+  const Wrapper = props => {
+    const [deleteFunc, deleteMeta] = useDelete(options);
+    return <C {...props} {...{ [funcName]: deleteFunc }} {...{ [metaName]: deleteMeta }} removeMutation={legacyError} />;
   };
   Wrapper.displayName = `withDelete${typeName}`;
   return Wrapper;
