+++ packages/vulcan-core/lib/modules/containers/update.js
@@ -30,58 +30,56 @@ Child Props:
 import React from 'react';
 import { useMutation } from '@apollo/client';
 import gql from 'graphql-tag';
-import {
-  updateClientTemplate,
-  extractCollectionInfo,
-  extractFragmentInfo,
-} from 'meteor/vulcan:lib';
+import { updateClientTemplate, extractCollectionInfo, extractFragmentInfo } from 'meteor/vulcan:lib';
+import { computeQueryVariables } from './variables';
 
-export const buildUpdateQuery = ({ typeName, fragmentName, fragment }) => (
+import { multiQueryUpdater } from './create';
+
+export const buildUpdateQuery = ({ typeName, fragmentName, fragment }) =>
   gql`
     ${updateClientTemplate({ typeName, fragmentName })}
     ${fragment}
-  `
-);
+  `;
 
-export const useUpdate = (options) => {
+export const useUpdate = options => {
   const { collectionName, collection } = extractCollectionInfo(options);
   const { fragmentName, fragment } = extractFragmentInfo(options, collectionName);
   const { mutationOptions = {} } = options;
 
   const typeName = collection.options.typeName;
-  const resolverName = `update${typeName}`;
+
   const query = buildUpdateQuery({ typeName, fragmentName, fragment });
 
   const [updateFunc, ...rest] = useMutation(query, {
     // see https://www.apollographql.com/docs/react/features/error-handling/#error-policies
     errorPolicy: 'all',
-    ...mutationOptions
-  }
-  );
+    update: multiQueryUpdater({ typeName, fragment, fragmentName, collection, resolverName: `update${typeName}` }),
+    ...mutationOptions,
+  });
 
-  const extendedUpdateFunc = {
-    [resolverName]: ({ data, selector }) => updateFunc({
-      variables: { data, selector },
-    })
-  }
-  return [extendedUpdateFunc[resolverName], ...rest];
+  const extendedUpdateFunc = ({ data, ...args }) => {
+    return updateFunc({
+      variables: {
+        data,
+        ...computeQueryVariables(options, args),
+      },
+    });
+  };
+  return [extendedUpdateFunc, ...rest];
 };
 
 export const withUpdate = options => C => {
   const { collection } = extractCollectionInfo(options);
   const typeName = collection.options.typeName;
   const funcName = `update${typeName}`;
+  const metaName = `update${typeName}Meta`;
 
   const legacyError = () => {
     throw new Error(`editMutation function has been removed. Use ${funcName} function instead.`);
   };
   const Wrapper = props => {
-    const [updateFunc] = useUpdate(options);
-    return <C
-      {...props}
-      {...{ [funcName]: updateFunc }}
-      editMutation={legacyError}
-    />;
+    const [updateFunc, updateMeta] = useUpdate(options);
+    return <C {...props} {...{ [funcName]: updateFunc }} {...{ [metaName]: updateMeta }} editMutation={legacyError} />;
   };
   Wrapper.displayName = `withUpdate${typeName}`;
   return Wrapper;
