+++ packages/vulcan-core/lib/modules/containers/update_old.js
@@ -30,71 +30,52 @@ Child Props:
 import React from 'react';
 import { useMutation } from '@apollo/client';
 import gql from 'graphql-tag';
-import {
-  updateClientTemplate,
-  extractCollectionInfo,
-  extractFragmentInfo,
-} from 'meteor/vulcan:lib';
-import { computeQueryVariables } from './variables';
+import { updateClientTemplate, extractCollectionInfo, extractFragmentInfo } from 'meteor/vulcan:lib';
 
-import { multiQueryUpdater } from './create';
-
-export const buildUpdateQuery = ({ typeName, fragmentName, fragment }) => (
+export const buildUpdateQuery = ({ typeName, fragmentName, fragment }) =>
   gql`
     ${updateClientTemplate({ typeName, fragmentName })}
     ${fragment}
-  `
-);
+  `;
 
-export const useUpdate2 = (options) => {
+export const useUpdateOld = options => {
   const { collectionName, collection } = extractCollectionInfo(options);
   const { fragmentName, fragment } = extractFragmentInfo(options, collectionName);
-  const {
-    mutationOptions = {}
-  } = options;
+  const { mutationOptions = {} } = options;
 
   const typeName = collection.options.typeName;
-
+  const resolverName = `update${typeName}`;
   const query = buildUpdateQuery({ typeName, fragmentName, fragment });
 
   const [updateFunc, ...rest] = useMutation(query, {
     // see https://www.apollographql.com/docs/react/features/error-handling/#error-policies
     errorPolicy: 'all',
-    update: multiQueryUpdater({ typeName, fragment, fragmentName, collection, resolverName: `update${typeName}` }),
-    ...mutationOptions
+    ...mutationOptions,
   });
 
-  const extendedUpdateFunc = ({ data, ...args }) => {
-    return updateFunc({
-      variables: {
-        data,
-        ...computeQueryVariables(options, args)
-      },
-    });
+  const extendedUpdateFunc = {
+    [resolverName]: ({ data, selector }) =>
+      updateFunc({
+        variables: { data, selector },
+      }),
   };
-  return [extendedUpdateFunc, ...rest];
+  return [extendedUpdateFunc[resolverName], ...rest];
 };
 
-export const withUpdate2 = options => C => {
+export const withUpdateOld = options => C => {
   const { collection } = extractCollectionInfo(options);
   const typeName = collection.options.typeName;
   const funcName = `update${typeName}`;
-  const metaName = `update${typeName}Meta`;
 
   const legacyError = () => {
     throw new Error(`editMutation function has been removed. Use ${funcName} function instead.`);
   };
   const Wrapper = props => {
-    const [updateFunc, updateMeta] = useUpdate2(options);
-    return <C
-      {...props}
-      {...{ [funcName]: updateFunc }}
-      {...{ [metaName]: updateMeta }}
-      editMutation={legacyError}
-    />;
+    const [updateFunc] = useUpdateOld(options);
+    return <C {...props} {...{ [funcName]: updateFunc }} editMutation={legacyError} />;
   };
-  Wrapper.displayName = `withUpdate${typeName}`;
+  Wrapper.displayName = `withUpdateOld${typeName}`;
   return Wrapper;
 };
 
-export default withUpdate2;
+export default withUpdateOld;
