+++ packages/vulcan-core/lib/modules/index.js
@@ -29,18 +29,18 @@ export * from './components/PaginatedList/index';
 
 export { default as withAccess } from './containers/withAccess.js';
 export { withMessages, useMessages } from './containers/withMessages.js';
-export { withMulti, useMulti } from './containers/multi.js';
-export { withMulti2, useMulti2 } from './containers/multi2.js';
-export { withSingle, useSingle } from './containers/single.js';
-export { withSingle2, useSingle2 } from './containers/single2.js';
-export { withCreate, useCreate } from './containers/create.js';
-export { withCreate2, useCreate2 } from './containers/create2.js';
-export { withUpdate, useUpdate } from './containers/update.js';
-export { withUpdate2, useUpdate2 } from './containers/update2.js';
-export { withUpsert, useUpsert } from './containers/upsert.js';
-export { withUpsert2, useUpsert2 } from './containers/upsert2.js';
-export { withDelete, useDelete } from './containers/delete.js';
-export { withDelete2, useDelete2 } from './containers/delete2.js';
+export { withMultiOld, useMultiOld } from './containers/multi_old';
+export { withMulti, useMulti } from './containers/multi';
+export { withSingleOld, useSingleOld } from './containers/single_old';
+export { withSingle, useSingle } from './containers/single';
+export { withCreateOld, useCreateOld } from './containers/create_old';
+export { withCreate, useCreate } from './containers/create';
+export { withUpdateOld, useUpdateOld } from './containers/update_old';
+export { withUpdate, useUpdate } from './containers/update';
+export { withUpsertOld, useUpsertOld } from './containers/upsert_old';
+export { withUpsert, useUpsert } from './containers/upsert';
+export { withDeleteOld, useDeleteOld } from './containers/delete_old';
+export { withDelete, useDelete } from './containers/delete';
 export { withCurrentUser, useCurrentUser } from './containers/currentUser.js';
 export { withMutation, useRegisteredMutation } from './containers/registeredMutation.js';
 export { withSiteData, useSiteData } from './containers/siteData.js';
