+++ packages/vulcan-core/test/index.js
@@ -1,7 +1,7 @@
 import './components.test';
+import './containers_old/queries.test';
+import './containers_old/mutations.test';
 import './containers/queries.test';
 import './containers/mutations.test';
-import './containers2/queries.test';
-import './containers2/mutations.test';
 import './withComponents.test';
 import './menu.test';
