+++ packages/vulcan-lib/lib/server/main.js
@@ -14,6 +14,8 @@ export * from './mutators.js';
 export * from './errors.js';
 export * from './default_resolvers.js';
 export * from './default_mutations.js';
+export * from './default_resolvers_old.js';
+export * from './default_mutations_old.js';
 // TODO: what to do with this?
 export * from './meteor_patch.js';
 //export * from './render_context.js';
